Minor refactor: break frozen module in small modules
