+++ .gitignore
@@ -107,6 +107,7 @@ celerybeat.pid
 env/
 venv/
 venv-pypy/
+venv-graalpython/
 ENV/
 env.bak/
 venv.bak/
