+++ gelidum/frozen/__init__.py
@@ -1,0 +1,4 @@
+from gelidum.frozen.frozen_class_creator import (
+    make_frozen_class, get_frozen_classes, clear_frozen_classes
+)
+from gelidum.frozen.frozen_base import FrozenBase
