+++ tests/gelidum_tests/test_freeze.py
@@ -14,7 +14,7 @@ from unittest.mock import patch
 from frozendict import frozendict
 from gelidum import FrozenException
 from gelidum import freeze
-from gelidum.frozen import FrozenBase, clear_frozen_classes
+from gelidum.frozen import FrozenBase, get_frozen_classes, clear_frozen_classes
 
 
 class TestFreeze(unittest.TestCase):
@@ -951,10 +951,6 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(Dummy.__module__, frozen_dummy3.get_gelidum_hot_class_module())
 
     def test_count_frozen_classes(self):
-        import gelidum.frozen
-
-        frozen_classes = getattr(gelidum.frozen, "__FROZEN_CLASSES")
-
         class Dummy(object):
             def __init__(self, attr: int):
                 self.attr = attr
@@ -965,8 +961,9 @@ class TestFreeze(unittest.TestCase):
         frozen_dummy1 = freeze(dummy1, on_freeze="copy")
         frozen_dummy2 = freeze(dummy2, on_freeze="copy")
         frozen_dummy3 = freeze(dummy3, on_freeze="copy")
+        frozen_classes = get_frozen_classes()
 
-        self.assertSetEqual({frozen_dummy1.__class__}, set(frozen_classes.values()))
+        self.assertSetEqual({frozen_dummy1.__class__}, frozen_classes)
         self.assertEqual(frozen_dummy1.__class__, frozen_dummy2.__class__, frozen_dummy3.__class__)
 
     def test_invalid_str_for_on_freeze_parameter(self):
